Consolidate TeamsActivity-related types into single file per language
Merge TeamsActivityBuilder, TeamsChannelData, TeamsConversation, and related
types (TenantInfo, ChannelInfo, TeamInfo, MeetingInfo, NotificationInfo) into
the main TeamsActivity file in each language:

- .NET: TeamsActivity.cs now contains all Teams types
- Node: teams-activity.ts now contains all Teams types
- Python: teams_activity.py now contains all Teams types

Deleted redundant files:
- dotnet: TeamsActivityBuilder.cs, TeamsChannelData.cs, TeamsConversation.cs
- node: teams-activity-builder.ts, teams-channel-data.ts, teams-conversation.ts
- python: teams_activity_builder.py, teams_channel_data.py, teams_conversation.py

All tests pass across all three languages.
