+++ node/packages/botas/src/index.ts
@@ -15,7 +15,4 @@ export * from './activity-type.js'
 export * from './turn-context.js'
 export * from './remove-mention-middleware.js'
 export * from './teams-activity.js'
-export * from './teams-activity-builder.js'
-export * from './teams-channel-data.js'
-export * from './teams-conversation.js'
 export * from './suggested-actions.js'
