--- python/packages/botas/src/botas/teams_activity_builder.py
@@ -1,4 +1,0 @@
-# Re-export TeamsActivityBuilder from teams_activity module
-from botas.teams_activity import TeamsActivityBuilder
-
-__all__ = ["TeamsActivityBuilder"]
